+++ src/stdlib/date.c
@@ -9,7 +9,7 @@
 #include "zc.h"
 #include <time.h>
 
-char *zbuild_rfc1123_date_string(long t, char *buf)
+char *zbuild_rfc1123_date_string(long long t, char *buf)
 {
     struct tm tmbuf;
     gmtime_r((time_t *)(&t), &tmbuf);
@@ -17,7 +17,7 @@ char *zbuild_rfc1123_date_string(long t, char *buf)
     return buf;
 }
 
-char *zbuild_rfc822_date_string(long t, char *buf)
+char *zbuild_rfc822_date_string(long long t, char *buf)
 {
     struct tm tmbuf;
     localtime_r((time_t *)&t, &tmbuf);
